extract ImageSource
